+++ includes/src/Usuarios/Memoria/usuarios.php
@@ -1,33 +1,7 @@
 <?php
 //require_once dirname(dirname(dirname(__DIR__))).'/configuracion.php';
-//require_once RUTA_VIEW .'/helpers/autorizacion.php';
+require_once RUTA_VIEW .'/helpers/autorizacion.php';
 
-function saludo()
-{
-    $html = '';
-
-    if (estaLogado()) {
-        $urlLogout = Utils::buildUrl('/logout.php');
-        $html = <<<EOS
-        Bienvenido, {$_SESSION['nombre']} <a href="{$urlLogout}">(salir)</a>
-        EOS;
-    } else {
-        $urlLogin = Utils::buildUrl('/login.php');
-        $html = <<<EOS
-        Usuario desconocido. <a href="{$urlLogin}">Login</a>
-        EOS;
-    }
-
-    return $html;
-}
-
-function logout()
-{
-    //Doble seguridad: unset + destroy
-    unset($_SESSION['idUsuario']);
-    unset($_SESSION['roles']);
-    unset($_SESSION['nombre']);
+class Usuario {
     
-    session_destroy();
-    session_start();
 }
