+++ includes/vistas/commun/layout.php
@@ -1,0 +1,21 @@
+
+<!DOCTYPE html>
+<html lang="en">
+<html>
+	<head>
+		<meta charset="UTF-8">
+		<meta name="viewport" content="width=device-width, initial-scale=1.0">
+		<title>Bocetos</title>
+		<link id="estilo" rel="stylesheet" type="text/css" href="csses/style.css" />
+	</head>
+	<body>
+		<div id="contenedor">
+		<?php require('cabecera.php'); ?>
+		<main>
+			<body>
+				
+			</body>
+		</main>
+		</div>
+	</body>
+</html>
