+++ includes/vistas/helpers/usuariosHelp.php
@@ -1,0 +1,33 @@
+<?php
+require_once 'autorizacion.php';
+
+function saludo()
+{
+    $html = '';
+
+    if (estaLogado()) {
+        $urlLogout = Utils::buildUrl('/logout.php');
+        $html = <<<EOS
+        Bienvenido, {$_SESSION['nombre']} <a href="{$urlLogout}">(salir)</a>
+        EOS;
+    } else {
+        $urlLogin = Utils::buildUrl('/login.php');
+        $html = <<<EOS
+        Usuario desconocido. <a href="{$urlLogin}">Login</a>
+        EOS;
+    }
+
+    return $html;
+}
+
+function logout()
+{
+    //Doble seguridad: unset + destroy
+    unset($_SESSION['idUsuario']);
+    unset($_SESSION['roles']);
+    unset($_SESSION['nombre']);
+    //unset(_SESSION['contaseña'])
+
+    session_destroy();
+    session_start();
+}
