+++ index.php
@@ -1,24 +1,23 @@
-<?php 
-require_once 'includes/configuracion.php';
-?>
 <!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Dungeon's Home</title>
-    <link id="estilo" rel="stylesheet" type="text/css" href="csses/style.css" />
-</head>
+<html lang="es">
+    <head>
+        <meta charset="UTF-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1.0">
+        <title>Miembros</title>
+        <link id="estilo" rel="stylesheet" type="text/css" href="csses/miembros.css" />
+    </head>
 
-<body>
-    <?php require  RUTA_VIEW . '/commun/cabecera.php'; ?>
-    <h1>INICIO</h1>
-    <h2>BIENVENIDO A LA HOGUERA</h2>
-    <img src="img/Dungeon's_Hub.png" alt="DH_web_icon" height= "512" width= "512">
-    <h2>Descansa y reponte. La campaña te espera</h2>
-    <p>Dungeon's hub es una herramienta web pensada para todos los amantes de los juegos de rol que te permitirá al usuario crear, 
-        aprender sobre DnD (Dungeons and Dragons) y crear personajes y mapas de las mazmorras para tus campañas.
-    </p>
-    <p>También podrás hablar con otra gente a través de los foros y incluso unirte a sus campañas. Y si no sabes que hace una raza o una
-        clase, también la podrás consultar en nuestra página para crear a tu personaje.
-    </p>
+    <body>
+    <?php require 'includes/vistas/commun/layout.php'; ?>
+        <h1>INICIO</h1>
+        <h2>BIENVENIDO A LA HOGUERA</h2>
+        <img src="img/Dungeon's_Hub.png" alt="DH_web_icon" height= "512" width= "512">
+        <h2>Descansa y reponte. La campaña te espera</h2>
+        <p>Dungeon's hub es una herramienta web pensada para todos los amantes de los juegos de rol que te permitirá al usuario crear, 
+            aprender sobre DnD (Dungeons and Dragons) y crear personajes y mapas de las mazmorras para tus campañas.
+        </p>
+        <p>También podrás hablar con otra gente a través de los foros y incluso unirte a sus campañas. Y si no sabes que hace una raza o una
+            clase, también la podrás consultar en nuestra página para crear a tu personaje.
+        </p>
+    </body>
+</html>
